pcShop/requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
celery==5.2.1
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
Deprecated==1.2.13
Django==3.2.9
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
kombu==5.2.2
MarkupSafe==2.0.1
packaging==21.3
Pillow==8.4.0
prompt-toolkit==3.0.23
PyJWT==2.3.0
pyparsing==3.0.6
pytz==2021.3
redis==4.0.2
requests==2.26.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
